requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
awsebcli==3.18.1
black==19.10b0
botocore==1.15.49
cement==2.8.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.8
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
docutils==0.15.2
future==0.16.0
gunicorn==20.0.4
idna==2.7
jmespath==0.10.0
mock==4.0.2
pathspec==0.5.9
Pillow==7.2.0
psycopg2==2.8.5
pypiwin32==223
python-dateutil==2.8.0
pytz==2020.1
pywin32==228
PyYAML==5.3.1
regex==2020.6.8
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.1
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.24.3
wcwidth==0.1.9
whitenoise==5.1.0
